requirements.txt
alembic==1.12.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
click==8.1.7
Flask==2.3.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
python-dotenv==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
